Replac utils
