+++ lock-protocol/src/spec/mod.rs
@@ -7,6 +7,4 @@ mod concrete_refines_range;
 pub mod concrete_tree_spec;
 mod range_refines_atomic;
 mod range_spec;
-pub mod test;
 pub mod utils;
-pub mod utils2;
