+++ lock-protocol/src/spec/utils.rs
@@ -1,22 +1,28 @@
+use std::intrinsics::offset;
+
 use builtin::*;
 use builtin_macros::*;
+use vstd::bytes;
 use vstd::prelude::*;
 use vstd::arithmetic::power::*;
 use vstd::arithmetic::div_mod::*;
 use vstd::arithmetic::mul::*;
+use vstd::set::*;
 use vstd_extra::prelude::*;
+
+use crate::mm::child;
 
 use super::common::*;
 
 verus! {
 
+broadcast use vstd_extra::seq_extra::group_forall_seq_lemmas;
+
 pub struct NodeHelper;
 
 impl NodeHelper {
-    /// depth starts from 0 (root) to 3 (leaf),
-    /// level starts from 4 (root) to 1 (leaf).
-    /// The mapping formula is `depth = 4 - level`, which reflects the inverse relationship
-    /// between `level` and `depth`.
+    /// depth starts from 0(root) to 3 (leaf),
+    /// level starts from 4(root) to 1 (leaf).
     pub open spec fn level_to_dep(level: nat) -> nat
         recommends
             1 <= level <= 4,
@@ -31,7 +37,7 @@ impl NodeHelper {
         (4 - dep) as nat
     }
 
-    pub proof fn lemma_level_to_dep(level: nat)
+    pub proof fn lemma_level_to_dep_bijective(level: nat)
         requires
             1 <= level <= 4,
         ensures
@@ -39,7 +45,7 @@ impl NodeHelper {
     {
     }
 
-    pub proof fn lemma_dep_to_level(dep: nat)
+    pub proof fn lemma_dep_to_level_bijective(dep: nat)
         requires
             0 <= dep < 4,
         ensures
@@ -47,46 +53,53 @@ impl NodeHelper {
     {
     }
 
-    pub open spec fn size_at_dep(dep: nat) -> nat
+    /// The number of nodes at a given level.
+    pub closed spec fn size_at_dep(dep: nat) -> nat
         recommends
             0 <= dep < 4,
     {
         pow(512, dep) as nat
     }
 
-    pub proof fn lemma_size_at_dep()
+    pub proof fn lemma_size_at_dep_table()
         ensures
             Self::size_at_dep(0) == 1,
             Self::size_at_dep(1) == 512,
             Self::size_at_dep(2) == 262144,
             Self::size_at_dep(3) == 134217728,
     {
+        assert(Self::size_at_dep(0) == 1) by (compute_only);
+        assert(Self::size_at_dep(1) == 512) by (compute_only);
+        assert(Self::size_at_dep(2) == 262144) by (compute_only);
+        assert(Self::size_at_dep(3) == 134217728) by (compute_only);
     }
 
-    /// Returns the total number of nodes in a full tree with a branching factor of 512,
-    /// up to and including the specified maximum depth (`max_dep`).
-    pub open spec fn tree_size_spec(max_dep: int) -> nat
+    /// Returns the size of the tree with nodes at most `max_dep` depth.
+    pub closed spec fn tree_size_spec(max_dep: int) -> nat
         recommends
             0 <= max_dep < 4,
         decreases max_dep,
+        when max_dep >= 0
     {
         let cur_dep = max_dep as nat;
         if max_dep == 0 {
-            Self::size_at_dep(cur_dep)
-        } else if max_dep > 0 {
-            Self::size_at_dep(cur_dep) + Self::tree_size_spec(max_dep - 1)
+            1
         } else {
-            arbitrary()
+            512 * Self::tree_size_spec(max_dep - 1) + 1
         }
     }
 
-    pub proof fn lemma_tree_size_spec()
+    pub proof fn lemma_tree_size_spec_table()
         ensures
             Self::tree_size_spec(0) == 1,
             Self::tree_size_spec(1) == 513,
             Self::tree_size_spec(2) == 262657,
             Self::tree_size_spec(3) == 134480385,
     {
+        assert(Self::tree_size_spec(0) == 1) by (compute_only);
+        assert(Self::tree_size_spec(1) == 513) by (compute_only);
+        assert(Self::tree_size_spec(2) == 262657) by (compute_only);
+        assert(Self::tree_size_spec(3) == 134480385) by (compute_only);
     }
 
     pub open spec fn tree_size_weighted_spec(k: int) -> nat
@@ -119,9 +132,14 @@ impl NodeHelper {
             Self::tree_size_relation_spec(2),
             Self::tree_size_relation_spec(3),
     {
-        Self::lemma_tree_size_spec();
+        Self::lemma_tree_size_spec_table();
+        assert(Self::tree_size_relation_spec(0)) by (compute_only);
+        assert(Self::tree_size_relation_spec(1)) by (compute_only);
+        assert(Self::tree_size_relation_spec(2)) by (compute_only);
+        assert(Self::tree_size_relation_spec(3)) by (compute_only);
     }
 
+    /// The total size of the tree with 4 levels.
     pub open spec fn total_size() -> nat {
         Self::tree_size_spec(3)
     }
@@ -138,27 +156,47 @@ impl NodeHelper {
         ensures
             Self::valid_nid(Self::root_id()),
     {
-        Self::lemma_tree_size_spec();
+        Self::lemma_tree_size_spec_table();
     }
 
-    pub open spec fn nid_to_trace_rec(nid: NodeId, cur_level: nat, cur_rt: NodeId) -> Seq<nat>
+    // Lemmas about trace.
+    // A trace is a sequence of offsets from the root to a node.
+    // The length of the trace is at most 3.
+    pub open spec fn valid_trace(trace: Seq<nat>) -> bool {
+        &&& 0 <= trace.len() < 4
+        &&& forall_seq_values(trace, |offset: nat| 0 <= offset < 512)
+    }
+
+    /// The set of all valid node ids.
+    pub open spec fn valid_nid_set() -> Set<NodeId> {
+        Set::new(|nid: NodeId| Self::valid_nid(nid))
+    }
+
+    /// The set of all valid traces.
+    pub open spec fn valid_trace_set() -> Set<Seq<nat>> {
+        Set::new(|trace: Seq<nat>| Self::valid_trace(trace))
+    }
+
+    /// Returns the trace from cur_rt to the node with id `nid`.
+    pub closed spec fn nid_to_trace_rec(nid: NodeId, cur_level: nat, cur_rt: NodeId) -> Seq<nat>
         decreases cur_level,
     {
         if cur_level == 0 {
-            Seq::empty()
+            seq![]
         } else {
-            let sz = Self::tree_size_spec(cur_level - 1);
-            let offset = ((nid - cur_rt - 1) / sz as int) as nat;
-            let new_rt = cur_rt + offset * sz + 1;
-            if new_rt == nid {
+            let child_size = Self::tree_size_spec(cur_level - 1);
+            let offset = ((nid - cur_rt - 1) / child_size as int) as nat;
+            let child_root = cur_rt + offset * child_size + 1;
+            if child_root == nid {
                 seq![offset]
             } else {
-                seq![offset].add(Self::nid_to_trace_rec(nid, (cur_level - 1) as nat, new_rt))
+                seq![offset].add(Self::nid_to_trace_rec(nid, (cur_level - 1) as nat, child_root))
             }
         }
     }
 
-    pub open spec fn nid_to_trace(nid: NodeId) -> Seq<nat>
+    /// Returns the trace from root to the node with id `nid`.
+    pub closed spec fn nid_to_trace(nid: NodeId) -> Seq<nat>
         recommends
             Self::valid_nid(nid),
     {
@@ -169,16 +207,13 @@ impl NodeHelper {
         }
     }
 
-    pub open spec fn valid_trace(trace: Seq<nat>) -> bool {
-        &&& 0 <= trace.len() < 4
-        &&& forall|i| #![trigger trace[i]] 0 <= i < trace.len() ==> 0 <= trace[i] < 512
-    }
-
-    pub open spec fn trace_to_nid(trace: Seq<nat>, cur_rt: NodeId, cur_level: int) -> NodeId
+    /// Returns the node id from the trace.
+    pub closed spec fn trace_to_nid_rec(trace: Seq<nat>, cur_rt: NodeId, cur_level: int) -> NodeId
         recommends
             Self::valid_trace(trace),
             Self::valid_nid(cur_rt),
             0 <= cur_level <= 3,
+            trace.len() <= cur_level,
     {
         trace.fold_left_alt(
             (cur_rt, cur_level as int),
@@ -192,14 +227,140 @@ impl NodeHelper {
         ).0
     }
 
-    pub proof fn lemma_trace_to_nid_inductive(trace: Seq<nat>, cur_rt: NodeId, cur_level: int)
+    /// Returns the node id from the trace starting from root.
+    pub closed spec fn trace_to_nid(trace: Seq<nat>) -> NodeId
+        recommends
+            Self::valid_trace(trace),
+    {
+        Self::trace_to_nid_rec(trace, 0, 3)
+    }
+
+    /// Returns the child node id from the trace and offset.
+    pub open spec fn child_nid_from_trace_offset(trace: Seq<nat>, offset: nat) -> NodeId
+        recommends
+            Self::valid_trace(trace),
+            0 <= offset < 512,
+            trace.len() < 3,
+    {
+        let nid1 = Self::trace_to_nid(trace);
+        let level = 3 - trace.len();
+        let sz = Self::tree_size_spec(level - 1);
+        nid1 + offset * sz + 1
+    }
+
+    /// Returns the child node id from the node id and offset.
+    pub open spec fn get_child(nid: NodeId, offset: nat) -> NodeId
+        recommends
+            Self::valid_nid(nid),
+            Self::nid_to_dep(nid) < 3,
+            0 <= offset < 512,
+    {
+        Self::trace_to_nid(Self::nid_to_trace(nid).push(offset))
+    }
+
+    /// Returns the parent node id from the node id.
+    /// If the node id is root, the result is arbitrary.
+    pub open spec fn get_parent(nid: NodeId) -> NodeId
+        recommends
+            Self::valid_nid(nid),
+            nid != Self::root_id(),
+    {
+        Self::trace_to_nid(Self::nid_to_trace(nid).drop_last())
+    }
+
+    /// Returns true if `ch` is a child of `pa (the trace of 'pa' is equal to the drooping last element trace of 'ch').
+    pub open spec fn is_child(pa: NodeId, ch: NodeId) -> bool
+        recommends
+            Self::valid_nid(pa),
+            Self::valid_nid(ch),
+    {
+        &&& Self::nid_to_trace(pa) == Self::nid_to_trace(ch).drop_last()
+        &&& ch != Self::root_id()
+    }
+
+    /// Returns the offset of the node id from its parent.
+    pub open spec fn get_offset(nid: NodeId) -> nat
+        recommends
+            Self::valid_nid(nid),
+            nid != Self::root_id(),
+    {
+        let trace = Self::nid_to_trace(nid);
+        trace.last()
+    }
+
+    pub open spec fn nid_to_dep(nid: NodeId) -> nat
+        recommends
+            Self::valid_nid(nid),
+    {
+        Self::nid_to_trace(nid).len()
+    }
+
+    pub open spec fn nid_to_level(nid: NodeId) -> nat
+        recommends
+            Self::valid_nid(nid),
+    {
+        Self::dep_to_level(Self::nid_to_dep(nid))
+    }
+
+    /// Returns the size of the subtree rooted at `nid`.
+    pub open spec fn sub_tree_size(nid: NodeId) -> nat
+        recommends
+            Self::valid_nid(nid),
+    {
+        let level = Self::nid_to_level(nid);
+        Self::tree_size_spec(level - 1)
+    }
+
+    /// The next node id outside the subtree rooted at `nid`.
+    pub open spec fn next_outside_subtree(nid: NodeId) -> NodeId
+        recommends
+            Self::valid_nid(nid),
+    {
+        nid + Self::sub_tree_size(nid)
+    }
+
+    /// Returns 'true' if 'nd' is in the subtree of 'rt' (the trace of 'rt' is a prefix of the trace of 'nd')
+    pub open spec fn in_subtree(rt: NodeId, nd: NodeId) -> bool
+        recommends
+            Self::valid_nid(rt),
+            Self::valid_nid(nd),
+    {
+        Self::nid_to_trace(rt).is_prefix_of(Self::nid_to_trace(nd))
+    }
+
+    pub open spec fn in_subtree_range(rt: NodeId, nd: NodeId) -> bool
+        recommends
+            Self::valid_nid(rt),
+    {
+        rt <= nd < Self::next_outside_subtree(rt)
+    }
+
+    /// Returns the NodeIds in the path from the trace.
+    pub open spec fn trace_to_tree_path(trace: Seq<nat>) -> Seq<NodeId>
+        recommends
+            Self::valid_trace(trace),
+        decreases trace.len(),
+    {
+        let ids = trace.map(|i: int, offset: nat| Self::trace_to_nid(trace.subrange(0, i + 1)));
+        seq![Self::root_id()] + ids
+    }
+
+    /// Returns the traces in the subtree of the given trace.
+    pub open spec fn get_subtree_traces(trace: Seq<nat>) -> Set<Seq<nat>>
+        recommends
+            Self::valid_trace(trace),
+    {
+        Self::valid_trace_set().filter(|subtree_trace| trace.is_prefix_of(subtree_trace))
+    }
+
+    proof fn lemma_trace_to_nid_rec_inductive(trace: Seq<nat>, cur_rt: NodeId, cur_level: int)
         requires
             Self::valid_trace(trace),
             Self::valid_nid(cur_rt),
             0 <= cur_level <= 3,
         ensures
             if trace.len() > 0 {
-                Self::trace_to_nid(trace, cur_rt, cur_level) == Self::trace_to_nid(
+                Self::trace_to_nid_rec(trace, cur_rt, cur_level) == Self::trace_to_nid_rec(
                     trace.subrange(1, trace.len() as int),
                     cur_rt + trace[0] * Self::tree_size_spec(cur_level - 1) + 1,
                     cur_level - 1,
@@ -241,7 +402,7 @@ impl NodeHelper {
         }
     }
 
-    pub proof fn lemma_trace_to_nid_inner(trace: Seq<nat>)
+    proof fn lemma_trace_to_nid_inner(trace: Seq<nat>)
         requires
             Self::valid_trace(trace),
         ensures
@@ -289,25 +450,17 @@ impl NodeHelper {
         }
     }
 
-    pub open spec fn trace_to_nid_increment_spec(trace: Seq<nat>, offset: nat) -> bool {
-        let nid1 = Self::trace_to_nid_from_root(trace);
-        let nid2 = Self::trace_to_nid_from_root(trace.push(offset));
-        let level = 3 - trace.len();
-        let sz = Self::tree_size_spec(level - 1);
-        nid2 == nid1 + offset * sz + 1
-    }
-
-    pub proof fn lemma_trace_to_nid_increment(trace: Seq<nat>, offset: nat)
+    /// `child_nid_from_trace_offset` correctly returns the child node id from the trace and offset,
+    /// i.e. `child_nid_from_trace_offset(trace, offset) == trace_to_nid(trace.push(offset))`.
+    pub proof fn lemma_child_nid_from_trace_offset_sound(trace: Seq<nat>, offset: nat)
         requires
             0 <= offset < 512,
-            Self::valid_trace(trace.push(offset)),
+            Self::valid_trace(trace),
         ensures
-            Self::trace_to_nid_increment_spec(trace, offset),
+            Self::child_nid_from_trace_offset(trace, offset) == Self::trace_to_nid(
+                trace.push(offset),
+            ),
     {
-        assert(Self::valid_trace(trace)) by {
-            assert(trace.is_prefix_of(trace.push(offset)));
-        };
-
         let func = |param: (NodeId, int), offset: nat|
             {
                 let nid = param.0;
@@ -329,67 +482,73 @@ impl NodeHelper {
         };
     }
 
+    /// The result of `trace_to_nid` is the same as splitting the trace into two parts
+    /// and calling `trace_to_nid_rec` on the second part.
     pub proof fn lemma_trace_to_nid_split(
         trace1: Seq<nat>,
         trace2: Seq<nat>,
         cur_rt: NodeId,
         cur_level: int,
-    ) -> (nid: NodeId)
+    )
         requires
             Self::valid_trace(trace1.add(trace2)),
-            cur_rt == Self::trace_to_nid_from_root(trace1),
+            cur_rt == Self::trace_to_nid(trace1),
             cur_level == Self::level_to_dep(trace1.len()) - 1,
         ensures
-            nid == Self::trace_to_nid(trace2, cur_rt, cur_level),
-            nid == Self::trace_to_nid_from_root(trace1.add(trace2)),
-        decreases trace2.len(),
+            Self::trace_to_nid(trace1.add(trace2)) == Self::trace_to_nid_rec(
+                trace2,
+                cur_rt,
+                cur_level,
+            ),
+        decreases
+                trace2.len(),
+    // Induction proof on the length of the trace2
+
     {
         if trace2.len() == 0 {
             assert(trace1.add(trace2) =~= trace1);
-
-            cur_rt
         } else {
+            // Induction step: use `lemma_child_nid_from_trace_offset_sound` to move the first element
+            // of `trace2` to the end of `trace1` and then use the inductive hypothesis on the rest of `trace2`.
             let new_trace1 = trace1.push(trace2[0]);
             let new_trace2 = trace2.subrange(1, trace2.len() as int);
             assert(new_trace1.add(new_trace2) =~= trace1.add(trace2));
             let new_rt = cur_rt + trace2[0] * Self::tree_size_spec(cur_level - 1) + 1;
-            assert(new_rt == Self::trace_to_nid_from_root(new_trace1)) by {
+            assert(new_rt == Self::trace_to_nid(new_trace1)) by {
                 assert(trace1.is_prefix_of(trace1.add(trace2)));
                 assert(trace2.is_suffix_of(trace1.add(trace2)));
-                Self::lemma_trace_to_nid_increment(trace1, trace2[0]);
+                Self::lemma_child_nid_from_trace_offset_sound(trace1, trace2[0]);
             };
             Self::lemma_trace_to_nid_split(new_trace1, new_trace2, new_rt, cur_level - 1)
         }
     }
 
-    pub open spec fn trace_to_nid_from_root(trace: Seq<nat>) -> NodeId
-        recommends
-            Self::valid_trace(trace),
-    {
-        Self::trace_to_nid(trace, 0, 3)
-    }
-
-    pub proof fn lemma_nid_to_trace_rec(nid: NodeId, cur_level: nat, cur_rt: NodeId) -> (trace: Seq<
-        nat,
-    >)
+    /// `nid_to_trace_rec` correcly returns a trace from `cur_rt` to the node id `nid`.
+    /// By applying `trace_to_nid_rec' to the trace produced by `nid_to_trace_rec`, we can
+    /// reconstruct the original node id.
+    pub proof fn lemma_nid_to_trace_rec_sound(nid: NodeId, cur_level: nat, cur_rt: NodeId)
         requires
             Self::valid_nid(nid),
             Self::valid_nid(cur_rt),
             0 <= cur_level <= 3,
             cur_rt < nid < cur_rt + Self::tree_size_spec(cur_level as int),
         ensures
-            trace =~= Self::nid_to_trace_rec(nid, cur_level, cur_rt),
-            trace.len() <= cur_level,
-            Self::valid_trace(trace),
-            nid == Self::trace_to_nid(trace, cur_rt, cur_level as int),
+            Self::nid_to_trace_rec(nid, cur_level, cur_rt).len() <= cur_level,
+            Self::valid_trace(Self::nid_to_trace_rec(nid, cur_level, cur_rt)),
+            nid == Self::trace_to_nid_rec(
+                Self::nid_to_trace_rec(nid, cur_level, cur_rt),
+                cur_rt,
+                cur_level as int,
+            ),
         decreases cur_level,
+    // Induction proof on cur_level
+
     {
         if cur_level == 0 {
-            Seq::empty()
         } else {
             assert(Self::tree_size_spec(cur_level - 1) * 512 + 1 == Self::tree_size_spec(
                 cur_level as int,
-            )) by { Self::lemma_tree_size_spec() };
+            )) by { Self::lemma_tree_size_spec_table() };
 
             let sz = Self::tree_size_spec(cur_level - 1);
             // Prove offset < 512
@@ -416,42 +575,61 @@ impl NodeHelper {
 
             if new_rt == nid {
                 let trace = seq![offset];
-                assert(Self::trace_to_nid(trace, cur_rt, cur_level as int) == nid) by {
-                    Self::lemma_trace_to_nid_inductive(trace, cur_rt, cur_level as int);
+                assert(Self::trace_to_nid_rec(trace, cur_rt, cur_level as int) == nid) by {
+                    Self::lemma_trace_to_nid_rec_inductive(trace, cur_rt, cur_level as int);
                 };
-                trace
             } else {
-                let _trace = Self::lemma_nid_to_trace_rec(nid, (cur_level - 1) as nat, new_rt);
+                Self::lemma_nid_to_trace_rec_sound(nid, (cur_level - 1) as nat, new_rt);
+                let _trace = Self::nid_to_trace_rec(nid, (cur_level - 1) as nat, new_rt);
                 let trace = seq![offset].add(_trace);
-                assert(Self::trace_to_nid(_trace, new_rt, cur_level - 1) == nid);
+                assert(Self::trace_to_nid_rec(_trace, new_rt, cur_level - 1) == nid);
                 assert(new_rt == cur_rt + offset * sz + 1);
-                assert(Self::trace_to_nid(trace, cur_rt, cur_level as int) == nid) by {
-                    Self::lemma_trace_to_nid_inductive(trace, cur_rt, cur_level as int);
+                assert(Self::trace_to_nid_rec(trace, cur_rt, cur_level as int) == nid) by {
+                    Self::lemma_trace_to_nid_rec_inductive(trace, cur_rt, cur_level as int);
                     assert(_trace =~= trace.subrange(1, trace.len() as int));
                 };
-                trace
             }
         }
     }
 
-    pub proof fn lemma_nid_to_trace(nid: NodeId) -> (trace: Seq<nat>)
+    /// `nid_to_trace` correctly returns a trace of the node id `nid` starting from the root.
+    /// By applying `nid_to_trace` to the trace produced by `nid_to_trace`, we can
+    /// reconstruct the original node id.
+    pub proof fn lemma_nid_to_trace_sound(nid: NodeId)
         requires
             Self::valid_nid(nid),
         ensures
-            trace =~= Self::nid_to_trace(nid),
-            Self::valid_trace(trace),
-            Self::trace_to_nid_from_root(trace) == nid,
+            Self::valid_trace(Self::nid_to_trace(nid)),
+            Self::trace_to_nid(Self::nid_to_trace(nid)) == nid,
     {
         if nid != Self::root_id() {
-            Self::lemma_nid_to_trace_rec(nid, 3, 0)
+            Self::lemma_nid_to_trace_rec_sound(nid, 3, 0)
         } else {
-            Seq::empty()
         }
     }
 
+    /// 'trace_to_nid` is the left inverse of `nid_to_trace` between the valid node id set and the valid trace set.
+    pub proof fn lemma_nid_to_trace_left_inverse()
+        ensures
+            left_inverse_on(
+                |nid| Self::nid_to_trace(nid),
+                |trace| Self::trace_to_nid(trace),
+                Self::valid_nid_set(),
+                Self::valid_trace_set(),
+            ),
+    {
+        assert forall|nid| #[trigger]
+            Self::valid_nid_set().contains(nid) implies Self::valid_trace_set().contains(
+            Self::nid_to_trace(nid),
+        ) && Self::trace_to_nid(Self::nid_to_trace(nid)) == nid by {
+            Self::lemma_nid_to_trace_sound(nid);
+        };
+    }
+
+    /// `trace_to_nid_rec` correctly returns a node id from the trace starting from `cur_rt`.
+    /// We can reconstruct the trace using `nid_to_trace_rec` with the trace given by `trace_to_nid_rec`.
     #[verifier::spinoff_prover]
-    pub proof fn lemma_trace_to_nid_rec(trace: Seq<nat>, cur_rt: NodeId, cur_level: int) -> (nid:
-        NodeId)
+    pub proof fn lemma_trace_to_nid_rec_sound(trace: Seq<nat>, cur_rt: NodeId, cur_level: int)
         requires
             Self::valid_trace(trace),
             Self::valid_nid(cur_rt),
@@ -459,16 +637,21 @@ impl NodeHelper {
             0 <= cur_level <= 3,
             trace.len() <= cur_level,
         ensures
-            nid == Self::trace_to_nid(trace, cur_rt, cur_level),
-            cur_rt <= nid < cur_rt + Self::tree_size_spec(cur_level),
-            nid == cur_rt <==> trace.len() == 0,
-            trace.len() != 0 ==> trace =~= Self::nid_to_trace_rec(nid, cur_level as nat, cur_rt),
-        decreases trace.len(),
+            cur_rt <= Self::trace_to_nid_rec(trace, cur_rt, cur_level) < cur_rt
+                + Self::tree_size_spec(cur_level),
+            Self::trace_to_nid_rec(trace, cur_rt, cur_level) == cur_rt <==> trace.len() == 0,
+            trace.len() != 0 ==> trace =~= Self::nid_to_trace_rec(
+                Self::trace_to_nid_rec(trace, cur_rt, cur_level),
+                cur_level as nat,
+                cur_rt,
+            ),
+        decreases
+                trace.len(),
+    // Induction proof on the length of the trace
+
     {
         if trace.len() == 0 {
-            assert(Self::tree_size_spec(cur_level) > 0) by { Self::lemma_tree_size_spec() };
-
-            cur_rt
+            assert(Self::tree_size_spec(cur_level) > 0) by { Self::lemma_tree_size_spec_table() };
         } else {
             let new_trace = trace.subrange(1, trace.len() as int);
             assert(new_trace.len() == trace.len() - 1);
@@ -480,7 +663,7 @@ impl NodeHelper {
             }) by {
                 assert(Self::tree_size_spec(cur_level - 1) * 512 + 1 == Self::tree_size_spec(
                     cur_level,
-                )) by { Self::lemma_tree_size_spec() };
+                )) by { Self::lemma_tree_size_spec_table() };
                 assert(Self::tree_size_spec(cur_level) - 1 - trace[0] * Self::tree_size_spec(
                     cur_level - 1,
                 ) == (512 - trace[0]) * Self::tree_size_spec(cur_level - 1)) by {
@@ -501,13 +684,15 @@ impl NodeHelper {
                         Self::tree_size_spec(cur_level - 1) as int,
                     )
                 };
-                assert(Self::tree_size_spec(cur_level - 1) > 0) by { Self::lemma_tree_size_spec() };
+                assert(Self::tree_size_spec(cur_level - 1) > 0) by {
+                    Self::lemma_tree_size_spec_table()
+                };
             };
 
             let new_level = cur_level - 1;
 
-            let nid = Self::lemma_trace_to_nid_rec(new_trace, new_rt, new_level);
-            assert(nid == Self::trace_to_nid(trace, cur_rt, cur_level));
+            Self::lemma_trace_to_nid_rec_sound(new_trace, new_rt, new_level);
+            let nid = Self::trace_to_nid_rec(trace, cur_rt, cur_level);
 
             let sz = Self::tree_size_spec(cur_level - 1);
             assert(new_rt <= nid < new_rt + sz);
@@ -519,7 +704,7 @@ impl NodeHelper {
             assert(new_rt + sz <= cur_rt + Self::tree_size_spec(cur_level)) by {
                 // tree_size_spec(cur_level) = 512*sz + 1
                 assert(Self::tree_size_spec(cur_level) == Self::tree_size_spec(cur_level - 1) * 512
-                    + 1) by { Self::lemma_tree_size_spec() };
+                    + 1) by { Self::lemma_tree_size_spec_table() };
 
                 // need: trace[0]*sz + sz <= 512*sz
                 assert(trace[0] < 512);
@@ -557,35 +742,70 @@ impl NodeHelper {
 
             let _trace = Self::nid_to_trace_rec(nid, cur_level as nat, cur_rt);
             assert(_trace =~= seq![offset].add(new_trace));
-
-            nid
         }
     }
 
-    pub proof fn lemma_trace_to_nid_from_root(trace: Seq<nat>) -> (nid: NodeId)
+    /// `trace_to_nid` correctly returns a node id from the trace starting from the root.
+    /// We can reconstruct the node id using `trace_to_nid` with the trace given by `nid_to_trace`.
+    pub proof fn lemma_trace_to_nid_sound(trace: Seq<nat>)
         requires
             Self::valid_trace(trace),
         ensures
-            nid == Self::trace_to_nid_from_root(trace),
-            Self::valid_nid(nid),
-            trace =~= Self::nid_to_trace(nid),
+            Self::valid_nid(Self::trace_to_nid(trace)),
+            trace =~= Self::nid_to_trace(Self::trace_to_nid(trace)),
     {
-        Self::lemma_tree_size_spec();
-        Self::lemma_trace_to_nid_rec(trace, 0, 3)
+        Self::lemma_tree_size_spec_table();
+        Self::lemma_trace_to_nid_rec_sound(trace, 0, 3)
     }
 
-    pub open spec fn nid_to_dep(nid: NodeId) -> nat
-        recommends
-            Self::valid_nid(nid),
+    /// `nid_to_trace` is the right inverse of `trace_to_nid` between the valid node id set and the valid trace set.
+    pub proof fn lemma_nid_to_trace_right_inverse()
+        ensures
+            right_inverse_on(
+                |trace| Self::nid_to_trace(trace),
+                |nid| Self::trace_to_nid(nid),
+                Self::valid_nid_set(),
+                Self::valid_trace_set(),
+            ),
     {
-        Self::nid_to_trace(nid).len()
+        assert forall|trace| #[trigger]
+            Self::valid_trace_set().contains(trace) implies Self::valid_nid_set().contains(
+            Self::trace_to_nid(trace),
+        ) && Self::nid_to_trace(Self::trace_to_nid(trace)) == trace by {
+            Self::lemma_trace_to_nid_sound(trace);
+        };
     }
 
-    pub open spec fn nid_to_level(nid: NodeId) -> nat
-        recommends
-            Self::valid_nid(nid),
+    /// The function `nid_to_trace` is bijective between the valid node id set and the valid trace set.
+    pub proof fn lemma_nid_to_trace_bijective()
+        ensures
+            bijective_on(
+                |nid| Self::nid_to_trace(nid),
+                Self::valid_nid_set(),
+                Self::valid_trace_set(),
+            ),
     {
-        Self::dep_to_level(Self::nid_to_dep(nid))
+        Self::lemma_nid_to_trace_left_inverse();
+        Self::lemma_nid_to_trace_right_inverse();
+    }
+
+    /// The function `trace_to_nid` is bijective between the valid trace set and the valid node id set.
+    pub proof fn lemma_trace_to_nid_bijective()
+        ensures
+            bijective_on(
+                |trace| Self::trace_to_nid(trace),
+                Self::valid_trace_set(),
+                Self::valid_nid_set(),
+            ),
+    {
+        Self::lemma_nid_to_trace_bijective();
+        Self::lemma_nid_to_trace_left_inverse();
+        lemma_inverse_of_bijection_is_bijective(
+            |nid| Self::nid_to_trace(nid),
+            |trace| Self::trace_to_nid(trace),
+            Self::valid_nid_set(),
+            Self::valid_trace_set(),
+        );
     }
 
     // Helper lemma: prove that the result length of nid_to_trace_rec does not exceed cur_level
@@ -595,6 +815,8 @@ impl NodeHelper {
         ensures
             Self::nid_to_trace_rec(nid, cur_level, cur_rt).len() <= cur_level,
         decreases cur_level,
+    // Induction proof on cur_level
+
     {
         if cur_level == 0 {
             assert(Self::nid_to_trace_rec(nid, 0, cur_rt).len() == 0);
@@ -614,6 +836,7 @@ impl NodeHelper {
         }
     }
 
+    /// `nid_to_depth` correctly returns a depth of 3 or less.
     pub proof fn lemma_nid_to_dep_le_3(nid: NodeId)
         requires
             Self::valid_nid(nid),
@@ -632,115 +855,79 @@ impl NodeHelper {
         }
     }
 
+    /// The relation between nid_to_level and nid_to_dep.
     pub proof fn lemma_level_dep_relation(nid: NodeId)
         requires
             Self::valid_nid(nid),
         ensures
             Self::nid_to_level(nid) == 4 - Self::nid_to_dep(nid),
     {
-        // Since nid_to_dep(nid) = nid_to_trace(nid).len()
-        // and nid_to_trace uses 3 levels of recursion at most, depth < 4
         assert(Self::nid_to_dep(nid) <= 3) by {
             Self::lemma_nid_to_dep_le_3(nid);
-        };  // Hence: dep ∈ {0,1,2,3}
+        };
     }
 
-    pub proof fn lemma_valid_level_to_node(nid: NodeId, level: nat)
+    /// A valid node's descendant's id is less than or equal to the total size.
+    pub proof fn lemma_valid_level_to_node(nid: NodeId)
         requires
             Self::valid_nid(nid),
-            level == Self::dep_to_level(Self::nid_to_dep(nid)),
         ensures
-            nid + Self::tree_size_spec(level - 1) <= Self::total_size(),
+            nid + Self::tree_size_spec(Self::nid_to_level(nid) - 1) <= Self::total_size(),
     {
-        let trace = Self::lemma_nid_to_trace(nid);
-        assert(nid == Self::trace_to_nid_from_root(trace)) by {
-            Self::lemma_nid_to_trace(nid);
-        };
+        let trace = Self::nid_to_trace(nid);
+        Self::lemma_nid_to_trace_sound(nid);
         assert(nid <= Self::trace_to_nid_upperbound_spec(trace.len())) by {
             Self::lemma_trace_to_nid_inner(trace);
         };
         assert(Self::trace_to_nid_upperbound_spec(trace.len()) == Self::tree_size_spec(3)
             - Self::tree_size_spec(3 - trace.len())) by {
-            Self::lemma_trace_to_nid_upperbound_spec(trace.len())
+            Self::lemma_trace_to_nid_upperbound_spec(trace.len());
         };
     }
 
-    pub open spec fn sub_tree_size(nid: NodeId) -> nat
-        recommends
-            Self::valid_nid(nid),
-    {
-        let dep = Self::nid_to_dep(nid);
-        let level = Self::dep_to_level(dep);
-        Self::tree_size_spec(level - 1)
-    }
-
+    /// A valid node's subtree size is at least 1.
     pub proof fn lemma_sub_tree_size_lowerbound(nid: NodeId)
         requires
             Self::valid_nid(nid),
         ensures
             Self::sub_tree_size(nid) >= 1,
     {
-        Self::lemma_tree_size_spec();
-        Self::lemma_nid_to_trace(nid);
+        Self::lemma_tree_size_spec_table();
+        Self::lemma_nid_to_trace_sound(nid);
     }
 
-    pub proof fn lemma_sub_tree_size_bounded(nid: NodeId)
+    /// A valid node plus its subtree size is less than or equal to the total size.
+    pub proof fn lemma_next_outside_subtree_bounded(nid: NodeId)
         requires
             Self::valid_nid(nid),
         ensures
-            nid + Self::sub_tree_size(nid) <= Self::total_size(),
+            Self::next_outside_subtree(nid) <= Self::total_size(),
     {
-        let level = Self::dep_to_level(Self::nid_to_dep(nid));
-        Self::lemma_valid_level_to_node(nid, level);
+        Self::lemma_valid_level_to_node(nid);
     }
 
-    pub open spec fn next_outside_subtree(nid: NodeId) -> NodeId
-        recommends
-            Self::valid_nid(nid),
-    {
-        nid + Self::sub_tree_size(nid)
-    }
-
-    pub proof fn lemma_next_outside_subtree_bounded(nid: NodeId) -> (nxt: NodeId)
-        requires
-            Self::valid_nid(nid),
-        ensures
-            nxt == Self::next_outside_subtree(nid),
-            nxt <= Self::total_size(),
-    {
-        Self::lemma_sub_tree_size_bounded(nid);
-        nid + Self::sub_tree_size(nid)
-    }
-
-    pub open spec fn in_subtree(rt: NodeId, nd: NodeId) -> bool
-        recommends
-            Self::valid_nid(rt),
-            Self::valid_nid(nd),
-    {
-        let sz = Self::sub_tree_size(rt);
-        rt <= nd < rt + sz
-    }
-
-    pub proof fn lemma_in_subtree_0(nid: NodeId)
+    /// A node is in its own subtree.
+    pub proof fn lemma_in_subtree_self(nid: NodeId)
         requires
             Self::valid_nid(nid),
         ensures
             Self::in_subtree(nid, nid),
     {
-        Self::lemma_sub_tree_size_lowerbound(nid);
     }
 
-    pub proof fn lemma_in_subtree(rt: NodeId, nd: NodeId)
+    /// A child is in the subtree of its parent.
+    pub proof fn lemma_is_child_implies_in_subtree(pa: NodeId, ch: NodeId)
         requires
-            Self::valid_nid(rt),
-            Self::valid_nid(nd),
-            Self::in_subtree(rt, nd),
+            Self::valid_nid(pa),
+            Self::valid_nid(ch),
+            Self::is_child(pa, ch),
         ensures
-            rt <= nd < Self::next_outside_subtree(rt),
+            Self::in_subtree(pa, ch),
+            Self::nid_to_dep(pa) + 1 == Self::nid_to_dep(ch),
     {
     }
 
-    pub proof fn lemma_in_subtree_bounded_is_child(pa: NodeId, ch: NodeId)
+    pub proof fn lemma_is_child_bound(pa: NodeId, ch: NodeId)
         requires
             Self::valid_nid(pa),
             Self::valid_nid(ch),
@@ -750,26 +937,15 @@ impl NodeHelper {
     {
         let dep_pa = Self::nid_to_dep(pa);
         let dep_ch = Self::nid_to_dep(ch);
-        assert(dep_ch == dep_pa + 1) by {
-            Self::lemma_is_child_properties(pa, ch);
-        };
-
         let sz_pa = Self::sub_tree_size(pa);
         let sz_ch = Self::sub_tree_size(ch);
-        Self::lemma_tree_size_spec();  // Apply tree size lemma to both nodes.
-
-        let level_pa = 4 - dep_pa;
-        let level_ch = 4 - dep_ch;
-
-        assert(level_ch == level_pa - 1);
-
+        Self::lemma_tree_size_spec_table();  // Apply tree size lemma to both nodes.
         // Verify trace properties.
-        let trace_pa = Self::lemma_nid_to_trace(pa);
-        let trace_ch = Self::lemma_nid_to_trace(ch);
-
-        // Set level bounds.
-        assert(1 <= level_ch <= 3);
-        assert(2 <= level_pa <= 4);
+        let trace_pa = Self::nid_to_trace(pa);
+        let trace_ch = Self::nid_to_trace(ch);
+        Self::lemma_nid_to_trace_sound(pa);
+        Self::lemma_nid_to_trace_sound(ch);
+        Self::lemma_is_child_implies_in_subtree(pa, ch);
 
         // Verify subtree containment.
         assert((ch as int) + (sz_ch as int) <= (pa as int) + (sz_pa as int)) by {
@@ -779,41 +955,24 @@ impl NodeHelper {
             // 2. If ch > pa, we need more detailed analysis
             if ch == pa {
             } else {
-                // Verify trace length relationship.
-                assert(trace_ch.len() == trace_pa.len() + 1);
-
                 // Prepare for increment lemma.
                 let offset = trace_ch.last();  // Get last element as offset.
-
                 // Get parent level.
                 let pa_level = Self::dep_to_level(dep_pa);
-
                 let sz = Self::tree_size_spec(pa_level - 2);
-
-                // Apply increment specification.
-                assert(Self::trace_to_nid_increment_spec(trace_pa, offset)) by {
-                    Self::lemma_trace_to_nid_increment(trace_pa, offset);
-                };
-
-                // The trace of ch is determined by its parent node's trace plus an offset
-                assert(ch != Self::root_id()) by {
-                    // A child node cannot be the root node, because it has depth > 0
-                    assert(dep_ch > 0);
-                    assert(Self::root_id() == 0);
-                    assert(ch > 0);
-                };
+                Self::lemma_child_nid_from_trace_offset_sound(trace_pa, offset);
 
                 // Final formula: ch = pa + offset * sz + 1, where sz = tree_size_spec(pa_level - 2).
                 assert(ch == pa + offset * sz + 1) by {
                     // By definition, get_parent uses trace.drop_last()
-                    assert(Self::get_parent(ch) == Self::trace_to_nid_from_root(
+                    assert(Self::get_parent(ch) == Self::trace_to_nid(
                         Self::nid_to_trace(ch).drop_last(),
                     ));
 
                     // Direct proof that trace_ch's prefix must be trace_pa
                     assert(trace_ch.drop_last() =~= trace_pa) by {
                         // Using uniqueness of trace to nid mapping
-                        Self::lemma_trace_to_nid_from_root(trace_ch.drop_last());
+                        Self::lemma_trace_to_nid_sound(trace_ch.drop_last());
                     }
 
                     // Now we prove that the last element of trace_ch is offset
@@ -822,14 +981,10 @@ impl NodeHelper {
                         assert(trace_ch =~= trace_pa.push(offset)) by {};
 
                         // Since trace_ch =~= trace_pa.push(offset), we know
-                        assert(Self::trace_to_nid_from_root(trace_ch)
-                            == Self::trace_to_nid_from_root(trace_pa.push(offset)));
+                        assert(Self::trace_to_nid(trace_ch) == Self::trace_to_nid(
+                            trace_pa.push(offset),
+                        ));
                     }
-
-                    // With these two facts, we can construct the full proof:
-                    // 1. trace_ch.drop_last() =~= trace_pa
-                    // 2. trace_ch.last() == offset
-                    // Therefore: trace_ch =~= trace_pa.push(offset)
                     assert(ch == pa + offset * sz + 1);
                 };
 
@@ -859,6 +1014,8 @@ impl NodeHelper {
         };
     }
 
+    // If 'nd' is in the subtree of 'rt', then `next_outside_subtree(nd)` (the next node id outside the subtree of 'nd') is less than or equal to
+    // `next_outside_subtree(rt)`.
     pub proof fn lemma_in_subtree_bounded(rt: NodeId, nd: NodeId)
         requires
             Self::valid_nid(rt),
@@ -866,241 +1023,531 @@ impl NodeHelper {
             Self::in_subtree(rt, nd),
         ensures
             Self::next_outside_subtree(nd) <= Self::next_outside_subtree(rt),
-        decreases Self::nid_to_dep(nd) - Self::nid_to_dep(rt),
-    {
-        // if rt != nd {
-        //     let pa = Self::get_parent(nd);
-        //     assert(Self::nid_to_dep(pa) + 1 == Self::nid_to_dep(nd));
-        //     Self::lemma_in_subtree_bounded(rt, pa);
-        //     Self::lemma_in_subtree_bounded_is_child(pa, nd);
-        // }
-        admit();
-    }
+        decreases
+                Self::nid_to_trace(nd).len() - Self::nid_to_trace(
+                    rt,
+                ).len(),
+    // Induction proof on the length of the trace between 'nd' and 'rt'
 
-    pub open spec fn is_child(pa: NodeId, ch: NodeId) -> bool
-        recommends
-            Self::valid_nid(pa),
-            Self::valid_nid(ch),
     {
-        pa == Self::get_parent(ch) && ch != Self::root_id()
-    }
-
-    pub open spec fn get_parent(nid: NodeId) -> NodeId
-        recommends
-            Self::valid_nid(nid),
-            nid != Self::root_id(),
-    {
-        if nid == Self::root_id() {
-            arbitrary()
+        if rt == nd {
         } else {
-            let trace = Self::nid_to_trace(nid);
-            Self::trace_to_nid_from_root(trace.drop_last())
+            // Induction step: use `lemma_is_child_bound` to prove the relationship between `nd` and its parent,
+            // and then use the induction hypothesis between `rt` and `nd`'s parent.
+            let parent = Self::get_parent(nd);
+            assert(Self::nid_to_trace(rt).len() < Self::nid_to_trace(nd).len()) by {
+                if Self::nid_to_trace(rt).len() == Self::nid_to_trace(nd).len() {
+                    assert(Self::nid_to_trace(rt) == Self::nid_to_trace(nd));
+                    Self::lemma_nid_to_trace_sound(rt);
+                    Self::lemma_nid_to_trace_sound(nd);
+                    assert(rt == nd);
+                }
+            }
+            Self::lemma_get_parent_sound(nd);
+            assert(Self::nid_to_trace(rt).is_prefix_of(Self::nid_to_trace(parent)));
+            Self::lemma_in_subtree_bounded(rt, parent);
+            Self::lemma_is_child_bound(parent, nd);
         }
     }
 
-    pub proof fn lemma_is_child_properties(pa: NodeId, ch: NodeId)
-        requires
-            Self::valid_nid(pa),
-            Self::valid_nid(ch),
-            Self::is_child(pa, ch),
-        ensures
-            Self::in_subtree(pa, ch),
-            Self::nid_to_dep(pa) + 1 == Self::nid_to_dep(ch),
-    {
-        let trace = Self::nid_to_trace(ch);
-        assert(Self::valid_trace(trace)) by {
-            Self::lemma_nid_to_trace(ch);
-        };
-        assert(pa == Self::trace_to_nid_from_root(trace.drop_last()));
-        let pa_trace = Self::nid_to_trace(pa);
-        assert(pa_trace =~= trace.drop_last()) by {
-            Self::lemma_trace_to_nid_from_root(trace.drop_last());
-        };
-        assert(pa_trace.len() + 1 == trace.len());
-        assert(Self::nid_to_dep(pa) + 1 == Self::nid_to_dep(ch));
-
-        let pa_level = Self::dep_to_level(pa_trace.len());
-        let nid = Self::lemma_trace_to_nid_split(pa_trace, seq![trace.last()], pa, pa_level - 1);
-        assert(nid == ch) by {
-            assert(pa_trace.add(seq![trace.last()]) =~= trace);
-            assert(nid == Self::trace_to_nid_from_root(trace));
-            Self::lemma_nid_to_trace(ch);
-        };
-        assert(Self::valid_nid(pa)) by {
-            Self::lemma_trace_to_nid_from_root(trace.drop_last());
-        };
-        assert(pa + Self::tree_size_spec(pa_level - 1) <= Self::total_size()) by {
-            Self::lemma_valid_level_to_node(pa, pa_level)
-        };
-        Self::lemma_trace_to_nid_rec(seq![trace.last()], pa, pa_level - 1);
-    }
-
-    pub proof fn lemma_parent_child(nid: NodeId)
+    /// `get_parent` correctly returns the parent of a node.
+    /// The result indeed satisfies `is_child(get_parent(nid), nid)`.
+    pub proof fn lemma_get_parent_sound(nid: NodeId)
         requires
             Self::valid_nid(nid),
             nid != Self::root_id(),
         ensures
+            Self::valid_nid(Self::get_parent(nid)),
             Self::is_child(Self::get_parent(nid), nid),
     {
+        Self::lemma_nid_to_trace_sound(nid);
+        Self::lemma_trace_to_nid_sound(Self::nid_to_trace(nid).drop_last());
     }
 
-    pub open spec fn get_offset(nid: NodeId) -> nat
-        recommends
-            Self::valid_nid(nid),
-            nid != Self::root_id(),
-    {
-        if nid == Self::root_id() {
-            arbitrary()
-        } else {
-            let trace = Self::nid_to_trace(nid);
-            trace.last()
-        }
-    }
-
-    pub proof fn lemma_get_offset(nid: NodeId) -> (offset: nat)
+    /// `get_offset` returns the offset in a correct range.
+    pub proof fn lemma_get_offset_sound(nid: NodeId)
         requires
             Self::valid_nid(nid),
             nid != Self::root_id(),
         ensures
-            offset == Self::get_offset(nid),
-            valid_pte_offset(offset),
+            0 <= Self::get_offset(nid) < 512,
     {
-        let trace = Self::lemma_nid_to_trace(nid);
-        trace.last()
+        Self::lemma_nid_to_trace_sound(nid);
     }
 
-    pub open spec fn get_child(nid: NodeId, offset: nat) -> NodeId
-        recommends
-            Self::valid_nid(nid),
-            Self::nid_to_dep(nid) < 3,
-            valid_pte_offset(offset),
-    {
-        let level = Self::dep_to_level(Self::nid_to_dep(nid));
-        let sz = Self::tree_size_spec(level - 2);
-        nid + offset * sz + 1
-    }
-
-    pub proof fn lemma_get_child(nid: NodeId, offset: nat)
+    /// `get_child` correctly returns the child of a node.
+    /// The result indeed satisfies `is_child(nid, get_child(nid, offset))`.
+    pub proof fn lemma_get_child_sound(nid: NodeId, offset: nat)
         requires
             Self::valid_nid(nid),
             Self::nid_to_dep(nid) < 3,
-            valid_pte_offset(offset),
+            0 <= offset < 512,
         ensures
             Self::valid_nid(Self::get_child(nid, offset)),
             nid == Self::get_parent(Self::get_child(nid, offset)),
             offset == Self::get_offset(Self::get_child(nid, offset)),
     {
-        // ==== 0. Prep the dep/level/size and discharge all `recommends` ====
-        let dep = Self::nid_to_dep(nid);
-        assert(0 <= dep && dep < 4) by {
-            // from nid_to_dep(nid) < 3
-        };
-        let level = Self::dep_to_level(dep);
-        assert(1 <= level && level <= 4) by {
-            // lemma_dep_to_level: level = 4 - dep, and 0 ≤ dep < 4
-            Self::lemma_dep_to_level(dep);
-        };
-        let sz_dep = (level as int) - 2;
-        assert(0 <= sz_dep && sz_dep < 4) by {
-            // level ∈ {2,3,4} ⇒ level–2 ∈ {0,1,2} ⊂ [0,4)
-        };
-        let sz: nat = Self::tree_size_spec(sz_dep);
-
-        // ==== 1. Show the child is within [0, total_size()) ====
-
-        let child = Self::get_child(nid, offset);
-        // From lemma_valid_level_to_node(nid, level):
-        Self::lemma_valid_level_to_node(nid, level);
-        //   ⇒ nid + tree_size_spec(level-1) ≤ total_size
-        // We need:
-        //   nid + offset*sz + 1 ≤ nid + tree_size_spec(level-1)
-        //
-        // First:  offset*sz + 1 ≤ (offset+1)*sz
-        assert(sz >= 1) by {
-            Self::lemma_tree_size_spec();
-        };
-        assert(offset * (sz as int) + 1 <= (offset + 1) * (sz as int)) by {
-            assert(((offset as int) + 1) * (sz as int) == (offset as int) * (sz as int) + (
-            sz as int)) by {
-                lemma_mul_is_distributive_add(sz as int, offset as int, 1);
-            };
-            assert(offset * sz + 1 <= offset * sz + sz);
-        };
-        // Next: (offset+1)*sz ≤ 512*sz
-        assert((offset + 1) * (sz as int) <= 512 * (sz as int)) by {
-            lemma_mul_inequality(offset as int + 1, 512, sz as int);
-        };
-        // Finally:     512*sz < tree_size_spec(level-1)
-        assert(512 * (sz as int) < (Self::tree_size_spec((level as int) - 1) as int)) by {
-            // tree_size_spec(level) = 512 * tree_size_spec(level-1) + 1
-            Self::lemma_tree_size_spec();
-        };
-        // Chain them
-        assert(nid + offset * (sz as int) + 1 <= nid + Self::tree_size_spec((level as int) - 1))
-            by {}
-        assert(nid + offset * (sz as int) + 1 < Self::total_size()) by {}
-
-        // ==== 2. Show `child` agrees with `trace_to_nid_from_root(trace.push(offset))` ====
-
-        // Extract the root‐based trace for `nid`
-        let trace = Self::lemma_nid_to_trace(nid);
-        // By lemma_trace_to_nid_increment:
-        Self::lemma_trace_to_nid_increment(trace, offset);
-        assert(Self::trace_to_nid_from_root(trace.push(offset)) == child) by {}
-
-        // ==== 3. Round‐trip through get_parent and get_offset ====
-
-        // get_parent(child) unfolds to trace_to_nid_from_root((trace.push(offset)).drop_last())
-        assert(child != Self::root_id()) by {
-            // child ≥ nid+1 > 0
-        };
-        assert(Self::valid_nid(child));
-        assert((trace.push(offset)).drop_last() =~= trace) by {}
-        assert(Self::valid_trace((trace.push(offset)).drop_last()));
-        assert(Self::valid_trace(trace));
-        assert(Self::nid_to_trace(child) == trace.push(offset)) by {
-            Self::lemma_trace_to_nid_from_root(trace.push(offset));
-        };
-        assert(Self::get_parent(child) == Self::trace_to_nid_from_root(trace));
-        // But trace_to_nid_from_root(trace) = nid
-        Self::lemma_trace_to_nid_from_root(trace);
-        assert(Self::get_parent(child) == nid) by {}
-
-        // Finally, get_offset(child) = trace.last() = offset
-        assert(Self::get_offset(child) == offset);
+        Self::lemma_nid_to_trace_sound(nid);
+        Self::lemma_trace_to_nid_sound(Self::nid_to_trace(nid).push(offset));
+        assert(Self::nid_to_trace(nid).push(offset).drop_last() =~= Self::nid_to_trace(nid));
     }
 
-    pub open spec fn is_not_leaf(nid: NodeId) -> bool
-        recommends
-            Self::valid_nid(nid),
-    {
-        Self::nid_to_dep(nid) < 3
-    }
-
-    pub proof fn lemma_valid_nid_set_finite(nid: NodeId)
+    /// The valid nid set's cardinality is `total_size`.
+    pub proof fn lemma_valid_nid_set_cardinality()
         ensures
-            Set::new(|nid| Self::valid_nid(nid)).finite(),
-            Set::new(|nid| Self::valid_nid(nid)).len() == Self::total_size(),
+            Self::valid_nid_set().finite(),
+            Self::valid_nid_set().len() == Self::total_size(),
     {
-        assert(Set::new(|nid| Self::valid_nid(nid)) =~= Set::new(
-            |nid| 0 <= nid < Self::total_size(),
-        ));
+        assert(Self::valid_nid_set() =~= Set::new(|nid| 0 <= nid < Self::total_size()));
         lemma_nat_range_finite(0, Self::total_size());
     }
 
-    pub open spec fn trace_to_tree_path(trace: Seq<nat>) -> Seq<NodeId>
-        recommends
-            Self::valid_trace(trace),
-        decreases trace.len(),
+    /// The valid trace set's cardinality is `total_size`.
+    pub proof fn lemma_valid_trace_set_cardinality()
+        ensures
+            Self::valid_trace_set().finite(),
+            Self::valid_trace_set().len() == Self::total_size(),
     {
-        if trace.len() == 0 {
-            seq![Self::root_id()]
+        Self::lemma_valid_nid_set_cardinality();
+        Self::lemma_nid_to_trace_bijective();
+        lemma_bijective_cardinality(
+            |nid| Self::nid_to_trace(nid),
+            Self::valid_nid_set(),
+            Self::valid_trace_set(),
+        );
+    }
+
+    /// `get_subtree_traces` returns a finite set of traces.
+    pub proof fn lemma_get_subtree_traces_finite(trace: Seq<nat>)
+        ensures
+            Self::get_subtree_traces(trace).finite(),
+    {
+        Self::lemma_valid_trace_set_cardinality();
+    }
+
+    /// `get_subtree_traces` contains the input trace.
+    pub proof fn lemma_get_subtree_traces_contains_self(trace: Seq<nat>)
+        requires
+            Self::valid_trace(trace),
+        ensures
+            Self::get_subtree_traces(trace).contains(trace),
+            Self::get_subtree_traces(trace).len() > 0,
+    {
+        Self::lemma_get_subtree_traces_finite(trace);
+        axiom_set_contains_len(Self::get_subtree_traces(trace), trace);
+    }
+
+    /// `get_subtree_traces` is injective
+    pub proof fn lemma_get_subtree_traces_injective()
+        ensures
+            injective_on(|trace| Self::get_subtree_traces(trace), Self::valid_trace_set()),
+    {
+        assert forall|x, y|
+            #![trigger Self::get_subtree_traces(x), Self::get_subtree_traces(y)]
+            Self::valid_trace_set().contains(x) && Self::valid_trace_set().contains(y)
+                && Self::get_subtree_traces(x) == Self::get_subtree_traces(y) implies x == y by {
+            Self::lemma_get_subtree_traces_contains_self(x);
+            Self::lemma_get_subtree_traces_contains_self(y);
+            let z = choose|z| Self::get_subtree_traces(x).contains(z);
+            lemma_prefix_of_common_sequence(x, y, z);
+        }
+
+    }
+
+    /// The cardinality of the subtree traces is equal to the tree size specification.
+    pub proof fn lemma_get_subtree_traces_cardinality(trace: Seq<nat>)
+        requires
+            Self::valid_trace(trace),
+        ensures
+            Self::get_subtree_traces(trace).len() == Self::tree_size_spec(3 - trace.len()),
+        decreases 3 - trace.len(),
+    {
+        let subtree_trace_set = Self::get_subtree_traces(trace);
+        assert(subtree_trace_set.contains(trace));
+        Self::lemma_tree_size_spec_table();
+        Self::lemma_valid_trace_set_cardinality();
+        if trace.len() == 3 {
+            assert(forall|subtree_trace| #[trigger]
+                subtree_trace_set.contains(subtree_trace) ==> subtree_trace =~= trace);
+            assert(subtree_trace_set.is_singleton()) by {
+                assert(forall|x, y|
+                    #![trigger subtree_trace_set.contains(x),subtree_trace_set.contains(y)]
+                    subtree_trace_set.contains(x) && subtree_trace_set.contains(y) ==> x == y);
+            }
+            subtree_trace_set.lemma_singleton_size();
         } else {
-            let path = Self::trace_to_tree_path(trace.drop_last());
-            let offset = trace.last();
-            let sz = Self::tree_size_spec(3 - trace.len());
-            let nid = path.last();
-            path.push(nid + offset * sz + 1)
+            // Split the set between the singleton `trace` and `child_trace_set`.
+            let f = |child_trace: Seq<nat>| child_trace.len() > trace.len();
+            let child_trace_set = subtree_trace_set.filter(f);
+            let trace_singleton_set = subtree_trace_set.filter(
+                |child_trace: Seq<nat>| !f(child_trace),
+            );
+            assert(trace_singleton_set.is_singleton()) by {
+                assert(trace_singleton_set.contains(trace));
+                assert forall|x, y|
+                    #![trigger trace_singleton_set.contains(x), trace_singleton_set.contains(y)]
+                    trace_singleton_set.contains(x) && trace_singleton_set.contains(y) implies x
+                    == y by {
+                    assert(x =~= trace);
+                    assert(y =~= trace);
+                }
+            }
+            assert(subtree_trace_set.len() == child_trace_set.len() + 1) by {
+                lemma_set_separation(subtree_trace_set, f);
+                trace_singleton_set.lemma_singleton_size();
+                trace_singleton_set.lemma_singleton_size();
+            }
+
+            // Split `child_trace_set` into sets of traces with different offsets.
+            let offset_set = Set::new(|i: nat| 0 <= i < 512);
+            assert(offset_set.len() == 512 && offset_set.finite()) by {
+                lemma_nat_range_finite(0, 512);
+            }
+            // The set of all child traces
+            let child_traces = offset_set.map(|offset| trace.push(offset));
+            assert(child_traces.len() == 512 && child_traces.finite()) by {
+                // Prove that push is injective
+                assert forall|t1: nat, t2: nat|
+                    #![trigger trace.push(t1), trace.push(t2)]
+                    trace.push(t1) == trace.push(t2) implies t1 == t2 by {
+                    assert(trace.push(t1).last() == trace.push(t2).last());
+                }
+                lemma_injective_implies_injective_on(|offset| trace.push(offset), offset_set);
+                lemma_injective_map_cardinality(
+                    |offset| trace.push(offset),
+                    offset_set,
+                    offset_set,
+                );
+            }
+            assert(child_traces <= Self::valid_trace_set());
+            // The set of each child trace's subtree traces set
+            let child_subtree_trace_set = child_traces.map(
+                |child_trace| Self::get_subtree_traces(child_trace),
+            );
+            // Use induction hypothesis here, prove that each child trace's subtree traces set has the size of tree with a
+            // height of 2 - trace.len()
+            assert forall|child_trace| #[trigger]
+                child_traces.contains(child_trace) implies Self::get_subtree_traces(
+                child_trace,
+            ).len() == Self::tree_size_spec(2 - trace.len()) by {
+                Self::lemma_get_subtree_traces_cardinality(child_trace);
+            };
+            assert(forall|child_subtree_trace| #[trigger]
+                child_subtree_trace_set.contains(child_subtree_trace) ==> child_subtree_trace.len()
+                    == Self::tree_size_spec(2 - trace.len()));
+            assert(child_subtree_trace_set.len() == 512 && child_subtree_trace_set.finite()) by {
+                Self::lemma_get_subtree_traces_injective();
+                lemma_injective_map_cardinality(
+                    |child_trace| Self::get_subtree_traces(child_trace),
+                    Self::valid_trace_set(),
+                    child_traces,
+                );
+            }
+
+            // Show that the child traces are disjoint
+            assert(pairwise_disjoint(child_subtree_trace_set));
+            // Show that the arbitrary union of the child_subtree_trace_set is equal to the child_trace_set
+            assert(child_trace_set =~= arbitrary_union(child_subtree_trace_set)) by {
+                assert forall|child_trace: Seq<nat>| #[trigger]
+                    child_trace_set.contains(child_trace) == arbitrary_union(
+                        child_subtree_trace_set,
+                    ).contains(child_trace) by {
+                    if child_trace_set.contains(child_trace) {
+                        assert(trace.is_prefix_of(child_trace));
+                        let trace_child = child_trace.subrange(0, (trace.len() + 1) as int);
+                        let offset = child_trace[trace.len() as int];
+                        assert(trace_child =~= trace.push(offset));
+                        assert(offset_set.contains(offset));
+                        // Definition of arbitrary union
+                        assert(child_traces.contains(trace_child));
+                        assert(child_subtree_trace_set.contains(
+                            Self::get_subtree_traces(trace_child),
+                        ));
+                    }
+                    if arbitrary_union(child_subtree_trace_set).contains(child_trace) {
+                        let child_subtree_trace = choose|t: Set<Seq<nat>>|
+                            child_subtree_trace_set.contains(t) && t.contains(child_trace);
+                        let trace_child = choose|t: Seq<nat>| #[trigger]
+                            child_traces.contains(t) && Self::get_subtree_traces(t)
+                                == child_subtree_trace;
+                        assert(trace.is_prefix_of(trace_child));
+                        assert(trace_child.is_prefix_of(child_trace));
+                        assert(trace.is_prefix_of(child_trace));
+                    }
+                }
+            }
+
+            assert(child_trace_set.len() == 512 * Self::tree_size_spec(2 - trace.len())) by {
+                lemma_arbitrary_union_cardinality_under_disjointness_same_length(
+                    child_subtree_trace_set,
+                    Self::tree_size_spec(2 - trace.len()),
+                );
+            }
+        }
+    }
+
+    /// The cardinality of nodes in subtree is `sub_tree_size`
+    pub proof fn lemma_in_subtree_cardinality(nid: NodeId)
+        requires
+            Self::valid_nid(nid),
+        ensures
+            Self::valid_nid_set().filter(|id| Self::in_subtree(nid, id)).len()
+                == Self::sub_tree_size(nid),
+    {
+        let nid_trace = Self::nid_to_trace(nid);
+        let subtree_traces = Self::get_subtree_traces(nid_trace);
+        let subtree_ids = Self::valid_nid_set().filter(|id| Self::in_subtree(nid, id));
+        Self::lemma_nid_to_trace_sound(nid);
+        Self::lemma_valid_nid_set_cardinality();
+        assert(subtree_traces.len() == Self::sub_tree_size(nid) && subtree_traces.finite()) by {
+            Self::lemma_get_subtree_traces_cardinality(nid_trace);
+            Self::lemma_get_subtree_traces_finite(nid_trace);
+        }
+        assert(bijective_on(|id| Self::nid_to_trace(id), subtree_ids, subtree_traces)) by {
+            Self::lemma_nid_to_trace_bijective();
+            // Prove subtree_traces =~= subtree_ids.map(nid_to_trace)
+            assert forall|trace| #[trigger]
+                subtree_traces.contains(trace) == subtree_ids.map(
+                    |id| Self::nid_to_trace(id),
+                ).contains(trace) by {
+                if subtree_traces.contains(trace) {
+                    let trace_id = Self::trace_to_nid(trace);
+                    assert(subtree_ids.contains(trace_id) && Self::nid_to_trace(trace_id) == trace)
+                        by {
+                        Self::lemma_trace_to_nid_sound(trace);
+                    }
+                }
+            }
+            assert(subtree_traces =~= subtree_ids.map(|id| Self::nid_to_trace(id)));
+
+        }
+        lemma_bijective_cardinality(|id| Self::nid_to_trace(id), subtree_ids, subtree_traces);
+    }
+
+    /// If a node id is in the subtree, than it falls in the range of [rt, next_outside_subtree(rt)).
+    proof fn lemma_in_subtree_implies_in_subtree_range(rt: NodeId, nd: NodeId)
+        requires
+            Self::valid_nid(rt),
+            Self::valid_nid(nd),
+            Self::in_subtree(rt, nd),
+        ensures
+            Self::in_subtree_range(rt, nd),
+    {
+        assert(rt <= nd < Self::next_outside_subtree(rt)) by {
+            assert(Self::next_outside_subtree(nd) <= Self::next_outside_subtree(rt)) by {
+                Self::lemma_in_subtree_bounded(rt, nd);
+            };
+            assert(nd < Self::next_outside_subtree(nd)) by {
+                Self::lemma_sub_tree_size_lowerbound(nd);
+            };
+            assert(rt <= nd) by {
+                let rt_trace = Self::nid_to_trace(rt);
+                let nd_trace = Self::nid_to_trace(nd);
+
+                // First prove that both are valid traces
+                assert(Self::valid_trace(rt_trace)) by {
+                    Self::lemma_nid_to_trace_sound(rt);
+                }
+                assert(Self::valid_trace(nd_trace)) by {
+                    Self::lemma_nid_to_trace_sound(nd);
+                }
+
+                // Since in_subtree(rt, nd), we know rt_trace is a prefix of nd_trace
+                assert(rt_trace.is_prefix_of(nd_trace));
+
+                if rt_trace.len() == 0 {
+                    // rt is root (0), so 0 <= nd is trivially true
+                } else {
+                    if rt == nd {
+                        // Trivial case: rt == nd implies rt <= nd
+                    } else {
+                        // DIRECT APPROACH: use lemma_trace_to_nid_rec_sound directly
+                        // When one trace is a proper prefix of another, the corresponding node
+                        // is an ancestor in the tree and has a smaller node ID
+                        // Get the suffix part of nd_trace after rt_trace
+                        let suffix = nd_trace.subrange(
+                            rt_trace.len() as int,
+                            nd_trace.len() as int,
+                        );
+
+                        // Validate the level parameter for dep_to_level
+                        assert(0 <= rt_trace.len() < 4) by {
+                            // Valid trace has length < 4
+                        }
+
+                        // We know cur_level needs to be 3 - suffix.len()
+                        let cur_level = 3 - rt_trace.len();
+
+                        // Prove that suffix is a valid trace
+                        assert(Self::valid_trace(suffix)) by {
+                            assert(suffix.len() < 4);
+                            assert(forall|i: int|
+                                #![trigger suffix[i]]
+                                0 <= i < suffix.len() ==> 0 <= suffix[i] < 512) by {
+                                assert(forall|i: int|
+                                    #![trigger nd_trace[i + rt_trace.len()]]
+                                    0 <= i < suffix.len() ==> 0 <= nd_trace[i + rt_trace.len()]
+                                        < 512);
+                            }
+                        }
+
+                        // Apply trace_to_nid_rec_sound directly using our key insights:
+                        // 1. rt = trace_to_nid(rt_trace)
+                        // 2. nd = trace_to_nid(rt_trace.add(suffix))
+                        // 3. Using lemma_trace_to_nid_rec_sound, we know:
+                        //    cur_rt <= trace_to_nid_rec(trace, cur_rt, level)
+
+                        assert(cur_level >= 0);
+                        assert(Self::valid_nid(rt));
+
+                        // Check if rt has enough space for the subtree
+                        assert(rt + Self::tree_size_spec(cur_level) <= Self::total_size()) by {
+                            Self::lemma_next_outside_subtree_bounded(rt);
+                        }
+
+                        // Show nd = trace_to_nid_rec(suffix, rt, cur_level)
+                        assert(nd == Self::trace_to_nid_rec(suffix, rt, cur_level)) by {
+                            // Show rt_trace.add(suffix) ~= nd_trace
+                            assert(rt_trace.add(suffix) =~= nd_trace) by {
+                                assert(rt_trace.is_prefix_of(nd_trace));
+                                assert(rt_trace =~= nd_trace.subrange(0, rt_trace.len() as int));
+                            }
+
+                            // First, prove that rt == trace_to_nid(rt_trace)
+                            assert(rt == Self::trace_to_nid(rt_trace)) by {
+                                Self::lemma_nid_to_trace_sound(rt);
+                            }
+
+                            // Instead of directly proving nd == trace_to_nid(rt_trace.add(suffix)),
+                            // we'll use a different approach that relies on the bijectivity of nid_to_trace and trace_to_nid
+
+                            // First establish that nid_to_trace(nd) =~= rt_trace.add(suffix)
+                            assert(Self::nid_to_trace(nd) =~= rt_trace.add(suffix)) by {
+                                assert(Self::nid_to_trace(nd) =~= nd_trace);  // By definition of nd
+                                assert(nd_trace =~= rt_trace.add(suffix));  // Already proved above
+                            }
+
+                            // Now use the bijectivity property: if nid_to_trace(nd) == trace, then nd == trace_to_nid(trace)
+                            Self::lemma_nid_to_trace_bijective();
+
+                            // Use the right inverse property: trace_to_nid(nid_to_trace(nd)) == nd
+                            assert(Self::trace_to_nid(Self::nid_to_trace(nd)) == nd) by {
+                                Self::lemma_nid_to_trace_sound(nd);
+                            }
+
+                            // Since nid_to_trace(nd) =~= rt_trace.add(suffix), and we have
+                            // trace_to_nid(nid_to_trace(nd)) == nd
+                            // we need to show that trace_to_nid is preserved by the equivalence relation
+                            assert(Self::trace_to_nid(Self::nid_to_trace(nd)) == Self::trace_to_nid(
+                                rt_trace.add(suffix),
+                            )) by {
+                                // We need to show that if t1 =~= t2, then trace_to_nid(t1) == trace_to_nid(t2)
+                                // We'll use lemma_trace_to_nid_sound to help with this
+                                let t1 = Self::nid_to_trace(nd);
+                                let t2 = rt_trace.add(suffix);
+                                assert(Self::valid_trace(t1));
+                                assert(Self::valid_trace(t2));
+
+                                // Apply sound lemma to both traces
+                                Self::lemma_trace_to_nid_sound(t1);
+                                Self::lemma_trace_to_nid_sound(t2);
+
+                                // Now we have:
+                                // t1 =~= nid_to_trace(trace_to_nid(t1))
+                                // t2 =~= nid_to_trace(trace_to_nid(t2))
+                                // And t1 =~= t2
+                                // Therefore nid_to_trace(trace_to_nid(t1)) =~= nid_to_trace(trace_to_nid(t2))
+                                // By injectivity of nid_to_trace, trace_to_nid(t1) == trace_to_nid(t2)
+                                assert(t1 =~= t2);
+                                assert(t1 =~= Self::nid_to_trace(Self::trace_to_nid(t1)));
+                                assert(t2 =~= Self::nid_to_trace(Self::trace_to_nid(t2)));
+
+                                // Now use bijectivity to conclude trace_to_nid(t1) == trace_to_nid(t2)
+                                Self::lemma_nid_to_trace_right_inverse();
+                            }
+
+                            // Now we can conclude nd == trace_to_nid(rt_trace.add(suffix))
+                            assert(nd == Self::trace_to_nid(rt_trace.add(suffix)));
+
+                            // Now use lemma_trace_to_nid_split to complete the proof
+                            Self::lemma_trace_to_nid_split(rt_trace, suffix, rt, cur_level);
+
+                            // By the ensures clause of lemma_trace_to_nid_split, we have:
+                            // trace_to_nid(rt_trace.add(suffix)) == trace_to_nid_rec(suffix, rt, cur_level)
+                            // Therefore, nd == trace_to_nid_rec(suffix, rt, cur_level)
+                        }
+
+                        // Now directly apply lemma_trace_to_nid_rec_sound
+                        assert(rt <= Self::trace_to_nid_rec(suffix, rt, cur_level)) by {
+                            Self::lemma_trace_to_nid_rec_sound(suffix, rt, cur_level);
+                        }
+                    }
+                }
+            };
+        };
+    }
+
+    /// If a node id is in the range [rt, next_outside_subtree(rt)), then it is in the subtree (given by trace prefix).
+    /// We prove this by showing that we can find exactly `sub_tree_size(rt)` node ids in the range
+    /// `rt <= nd < next_outside_subtree(rt)` that is in the subtree, so every node id in the range
+    /// is in the subtree.
+    proof fn lemma_in_subtree_range_implies_in_subtree(rt: NodeId, nd: NodeId)
+        requires
+            Self::valid_nid(rt),
+            rt <= nd < Self::next_outside_subtree(rt),
+        ensures
+            Self::in_subtree(rt, nd),
+    {
+        // Showing every node id in the range is valid and the range is `Self::sub_tree_size(rt)`
+        let nid_set = Set::new(|nid: nat| rt <= nid < Self::next_outside_subtree(rt));
+        assert(nid_set.len() == Self::sub_tree_size(rt) && nid_set.finite()) by {
+            lemma_nat_range_finite(rt, Self::next_outside_subtree(rt));
+        }
+        assert(nid_set =~= nid_set.filter(|id| Self::valid_nid(id))) by {
+            assert forall|id| #[trigger] nid_set.contains(id) implies Self::valid_nid(id) by {
+                Self::lemma_next_outside_subtree_bounded(rt);
+            }
+        }
+        assert(nid_set =~= Self::valid_nid_set().filter(
+            |id| rt <= id < Self::next_outside_subtree(rt),
+        ));
+
+        // Shhowing there are exactly `sub_tree_size(rt)` node ids that are in the subtree
+        let child_set = Self::valid_nid_set().filter(|id| Self::in_subtree(rt, id));
+        assert(child_set.len() == Self::sub_tree_size(rt)) by {
+            Self::lemma_in_subtree_cardinality(rt);
+        };
+        // Every node id in the subtree falls in the range
+        assert(child_set =~= child_set.filter(|id| rt <= id < Self::next_outside_subtree(rt))) by {
+            assert forall|id| #[trigger] child_set.contains(id) implies rt <= id
+                < Self::next_outside_subtree(rt) by {
+                Self::lemma_in_subtree_implies_in_subtree_range(rt, id);
+            }
+        }
+        // So we can find exactly `sub_tree_size(rt)` node ids in the range that are in the subtree
+        assert(child_set =~= nid_set.filter(|id| Self::in_subtree(rt, id)));
+        assert(nid_set.len() == nid_set.filter(|id| Self::in_subtree(rt, id)).len());
+        // Therefore, every node id in the range is in the subtree
+        assert(nid_set == nid_set.filter(|id| Self::in_subtree(rt, id))) by {
+            lemma_filter_len_unchanged_implies_equal(nid_set, |id| Self::in_subtree(rt, id));
+        }
+        assert(nid_set.contains(nd));
+        assert(Self::in_subtree(rt, nd));
+    }
+
+    /// 'in_subtree' is equivalent to 'in_subtree_range' (nd in [rt, next_outside_subtree(rt)))
+    pub proof fn lemma_in_subtree_iff_in_subtree_range(rt: NodeId, nd: NodeId)
+        requires
+            Self::valid_nid(rt),
+        ensures
+            (Self::in_subtree(rt, nd) && Self::valid_nid(nd)) <==> Self::in_subtree_range(rt, nd),
+    {
+        if Self::in_subtree(rt, nd) && Self::valid_nid(nd) {
+            Self::lemma_in_subtree_implies_in_subtree_range(rt, nd);
+        }
+        if rt <= nd < Self::next_outside_subtree(rt) {
+            Self::lemma_next_outside_subtree_bounded(rt);
+            Self::lemma_in_subtree_range_implies_in_subtree(rt, nd);
         }
     }
 }
